+++ HISTORY.md
@@ -1,3 +1,8 @@
+## 1.0.4
+
+* documentation for deleting a collection, improved overall documentation
+* import Firestore data: support for merging with existing data
+
 ## 1.0.3
 
 * export Firebase Authentication